backend/pyproject.toml
[project]
name = "pixelcv-backend"
version = "2.0.0"
description = "Backend FastAPI para PixelCV - Sistema de CVs con comunidad y gamificación"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "pydantic-settings",
  "sqlalchemy>=2.0",
  "passlib[bcrypt]",
  "python-jose[cryptography]",
  "python-multipart",
  "pyyaml",
  "requests",
  "email-validator",
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "pytest-asyncio",
  "httpx",
]
